Update cleanup blob script
script processes a Max Restore Point. All files with lastmodified
more recent than ResotrePoint are omitted for processing. Trn and
Full Diffs required to restore to the Max restore point are added
to a set for keeping. Any items with lastmodified date older than
Max Restore point that have not been added to the "keep" set are
then deleted